# backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEERON API",
    "description": "Precision Aquaculture Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/dashboard/overview": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Dashboard Overview",
        "description": "Retrieves the executive overview metrics for a farm's Home Command Center.",
        "operationId": "get_dashboard_overview_api_v1_dashboard_overview_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_DashboardOverviewResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboard/trends": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Dashboard Trends",
        "description": "Returns risk trends over time for the farm.",
        "operationId": "get_dashboard_trends_api_v1_dashboard_trends_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_DashboardTrendsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/recommendations": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Dashboard Recommendations",
        "description": "Retrieves the most recent pending recommendations for the farm.",
        "operationId": "get_dashboard_recommendations_api_v1_dashboard_recommendations_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_RecommendationResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/farms": {
      "get": {
        "tags": [
          "Farms"
        ],
        "summary": "List Farms",
        "description": "Returns a list of all geographical farms registered in the platform.",
        "operationId": "list_farms_api_v1_farms_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_FarmResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/farms/{farm_id}": {
      "get": {
        "tags": [
          "Farms"
        ],
        "summary": "Get Farm Detail",
        "description": "Retrieves detailed properties for a single farm.",
        "operationId": "get_farm_detail_api_v1_farms__farm_id__get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_FarmResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tanks/{tank_id}": {
      "get": {
        "tags": [
          "Tanks"
        ],
        "summary": "Get Tank Detail",
        "description": "Retrieves static details and parameters for a specific tank.",
        "operationId": "get_tank_detail_api_v1_tanks__tank_id__get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_TankDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "Recommendations"
        ],
        "summary": "List Recommendations",
        "description": "Retrieves all recommendations, optionally filtered by tank_id, farm_id, or status.",
        "operationId": "list_recommendations_api_v1_recommendations_get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tank Id"
            }
          },
          {
            "name": "farm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farm Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "pending",
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_RecommendationResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/insights/dashboard": {
      "get": {
        "tags": [
          "Insights"
        ],
        "summary": "Get Dashboard Insights",
        "description": "Computes and returns a list of transient AI insights tailored for the dashboard.",
        "operationId": "get_dashboard_insights_api_v1_insights_dashboard_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_AiInsightResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/tanks/{tank_id}": {
      "get": {
        "tags": [
          "Insights"
        ],
        "summary": "Get Tank Insights",
        "description": "Computes and returns a list of transient AI insights for a specific tank.",
        "operationId": "get_tank_insights_api_v1_insights_tanks__tank_id__get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_AiInsightResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiInsightResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tankId": {
            "type": "string",
            "format": "uuid",
            "title": "Tankid"
          },
          "priority": {
            "type": "string",
            "maxLength": 50,
            "title": "Priority"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "title": {
            "type": "string",
            "maxLength": 150,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tankId",
          "priority",
          "confidence",
          "title",
          "description",
          "generatedAt"
        ],
        "title": "AiInsightResponse"
      },
...
      "BaseResponse_DashboardOverviewResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/DashboardOverviewResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[DashboardOverviewResponse]"
      },
      "BaseResponse_DashboardTrendsResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/DashboardTrendsResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[DashboardTrendsResponse]"
      },
...
      "BaseResponse_FarmResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/FarmResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[FarmResponse]"
      },
...
      "BaseResponse_List_AiInsightResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/AiInsightResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[AiInsightResponse]]"
      },
...
      "BaseResponse_List_FarmResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/FarmResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[FarmResponse]]"
      },
...
      "BaseResponse_List_RecommendationResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[RecommendationResponse]]"
      },
...
      "BaseResponse_TankDetailResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/TankDetailResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[TankDetailResponse]"
      },
...
      "DashboardOverviewResponse": {
        "properties": {
          "farmHealthScore": {
            "$ref": "#/components/schemas/FarmHealthScore"
          },
          "activeAlertsCount": {
            "type": "integer",
            "title": "Activealertscount"
          },
          "activeRecommendationsCount": {
            "type": "integer",
            "title": "Activerecommendationscount"
          },
          "averagePsi": {
            "type": "number",
            "title": "Averagepsi"
          },
          "waterQualitySummary": {
            "$ref": "#/components/schemas/WaterQualitySummary"
          },
          "zoneOverview": {
            "items": {
              "$ref": "#/components/schemas/DashboardZoneOverview"
            },
            "type": "array",
            "title": "Zoneoverview"
          },
          "recentInsights": {
            "items": {
              "$ref": "#/components/schemas/DashboardRecentInsight"
            },
            "type": "array",
            "title": "Recentinsights"
          }
        },
        "type": "object",
        "required": [
          "farmHealthScore",
          "activeAlertsCount",
          "activeRecommendationsCount",
          "averagePsi",
          "waterQualitySummary",
          "zoneOverview",
          "recentInsights"
        ],
        "title": "DashboardOverviewResponse"
      },
...
      "DashboardTrendsResponse": {
        "properties": {
          "farm_id": {
            "type": "string",
            "format": "uuid",
            "title": "Farm Id"
          },
          "trends": {
            "items": {
              "$ref": "#/components/schemas/RiskTrendPoint"
            },
            "type": "array",
            "title": "Trends"
          }
        },
        "type": "object",
        "required": [
          "farm_id",
          "trends"
        ],
        "title": "DashboardTrendsResponse"
      },
...
      "FarmHealthScore": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "trendPercent": {
            "type": "number",
            "title": "Trendpercent"
          },
          "classification": {
            "type": "string",
            "title": "Classification"
          }
        },
        "type": "object",
        "required": [
          "score",
          "trendPercent",
          "classification"
        ],
        "title": "FarmHealthScore"
      },
      "FarmResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "timezone": {
            "type": "string",
            "maxLength": 50,
            "title": "Timezone"
          },
          "carryingCapacityKg": {
            "type": "number",
            "title": "Carryingcapacitykg"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "latitude",
          "longitude",
          "timezone",
          "carryingCapacityKg"
        ],
        "title": "FarmResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tankId": {
            "type": "string",
            "format": "uuid",
            "title": "Tankid"
          },
          "action": {
            "type": "string",
            "maxLength": 200,
            "title": "Action"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "priority": {
            "type": "string",
            "maxLength": 50,
            "title": "Priority"
          },
          "expectedOutcome": {
            "type": "string",
            "title": "Expectedoutcome"
          },
          "status": {
            "type": "string",
            "maxLength": 50,
            "title": "Status"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tankId",
          "action",
          "confidence",
          "priority",
          "expectedOutcome",
          "status",
          "generatedAt"
        ],
        "title": "RecommendationResponse"
      },
...
      "TankDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "volumeM3": {
            "type": "number",
            "title": "Volumem3"
          },
          "maxBiomassKg": {
            "type": "number",
            "title": "Maxbiomasskg"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "digitalTwinConfig": {
            "additionalProperties": true,
            "type": "object",
            "title": "Digitaltwinconfig"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "volumeM3",
          "maxBiomassKg",
          "species",
          "digitalTwinConfig",
          "created_at"
        ],
        "title": "TankDetailResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}